requirements.txt
streamlit>=1.29.0
langchain>=0.3.0
langchain-core>=0.3.0
langchain-community>=0.3.0
langchain-google-genai>=2.0.0
google-generativeai>=0.8.0
langgraph>=0.2.0
qdrant-client>=1.12.0
pypdf>=4.0.0
duckduckgo-search>=6.0.0
python-dotenv
